requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
httpx==0.28.1
pydantic==2.10.3
pydantic-settings==2.6.1
chromadb==0.5.23
google-generativeai==0.8.3
pypdf==5.1.0